clean up the code and updated the ui
