+++ babel.config.js
@@ -9,8 +9,6 @@ module.exports = function (api) {
         {
           moduleName: "@env",
           path: ".env",
-          blacklist: null,
-          whitelist: null,
           safe: false,
           allowUndefined: true,
         },
