+++ package.json
@@ -19,6 +19,7 @@
     "@react-navigation/drawer": "^6.4.1",
     "@react-navigation/native": "^6.0.10",
     "@react-navigation/stack": "^6.2.1",
+    "@reduxjs/toolkit": "^1.9.1",
     "expo": "~45.0.0",
     "expo-font": "~10.1.0",
     "expo-linear-gradient": "~11.3.0",
@@ -38,6 +39,8 @@
     "react-native-safe-area-context": "4.2.4",
     "react-native-screens": "~3.11.1",
     "react-native-web": "0.17.7",
+    "react-query": "^3.39.2",
+    "react-redux": "^8.0.5",
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
