+++ src/components/Recipe/styles.js
@@ -1,12 +1,5 @@
 import styled from "styled-components/native";
-import { Text } from "../../../components/text";
-import { Feather } from "@expo/vector-icons";
-import { colors } from "../../../infrastructure/theme/colors";
-import { Button, Card, Searchbar } from "react-native-paper";
-import { Dimensions, ScrollView } from "react-native";
-
-const DEVICE_WIDTH = Dimensions.get("window").width;
-const DEVICE_HEIGHT = Dimensions.get("window").height;
+import { DEVICE_HEIGHT, DEVICE_WIDTH } from "../Header/styles";
 
 export const HeaderRow = styled.View`
   flex-direction: row;
@@ -18,8 +11,7 @@ export const HeaderRow = styled.View`
 export const RecipeImage = styled.View`
   width: ${(90 / 100) * DEVICE_WIDTH}px;
   height: ${(30 / 100) * DEVICE_HEIGHT}px;
-  margin: 0 auto;
-  margin-bottom: 15px;
+  margin: 10px auto;
   border-radius: 20px;
   overflow: hidden;
   elevation: 1;
