+++ src/components/safearea.js
@@ -4,4 +4,4 @@ import styled from "styled-components/native";
 export const SafeArea = styled(SafeAreaView)`
   flex: 1;
   ${StatusBar.currentHeight && `margin-top: ${StatusBar.currentHeight}px`};
-`;
+  `;
