--- src/features/home/screens/req.rest
@@ -1,1 +1,0 @@
-GET https://api.edamam.com/api/recipes/v2?type=public&q=chicken&app_id=49895665&app_key=2facef1175704a26f43a929745bb17f8
