--- src/infrastructure/theme/scheme.css
@@ -1,30 +1,0 @@
-.brand{
-    color:
-    #7D4A4A, 
-    #7d4a4ad2, 
-    #26262678,;
-}
-.ui {
-    color: 
-    #262626,
-    #757575,
-    #F1F1F1, 
-    #FFFFFF,
-    #DEDEDE,
-    #D0421B,
-    #138000,
-  }
-  .bg {
-    color: 
-    #FFFFFF, 
-    #2E2D40,
-  }
-  .text {
-    color: 
-    #7D4A4A,
-    #262626, 
-    #9C9C9C,
-    #FFFFFF, 
-    #D0421B, 
-    #138000,
-  }
