+++ src/navigation/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { AppNavigator } from "./App";
+
+export const Navigation = () => {
+  return <AppNavigator />;
+};
+
+// export {AppNavigator as Navigation} from './app.navigator';
